Shared vit_nwtc.h + refactor translations + integrate ComputeUA360_AttachedFlow
Create vit_nwtc.h with C++ translations of NWTC_Library utilities:
- Constants: Pi, TwoPi, PiBy2, D2R, R2D
- EqualRealNos (epsilon comparison, was inlined in 4 files)
- fZeros (zero-crossing finder, new translation)
- InterpExtrapStp (linear interpolation with stepping search, new)

Refactor 6 existing translations to use the shared header, removing
duplicate inline implementations. All 9 functions re-verified against
golden kernels (9/9 IDENTICAL).

Integrate ComputeUA360_AttachedFlow (10th AirfoilInfo function):
- 20/20 IDENTICAL kernel verification
- Uses fZeros, InterpExtrapStp, EqualRealNos from vit_nwtc.h
- Calls compute_iloweriupper_c via callee bridge
- --reverse-copy for 9 modified UA_BL scalar fields
- 16/16 baselines pass
- integrate_all.sh: 10/10 clean->integrated cycle

+++ kernel/FindBoundingTables/vit_types.h
@@ -127,8 +127,8 @@ typedef struct {
     double UserProp;
     double Re;
     int NumAlf;
-    int ConstData;
-    int InclUAdata;
+    int8_t ConstData;
+    int8_t InclUAdata;
     afi_ua_bl_type_t UA_BL;
 
     // --- ALLOCATABLE fields (pointer + size) ---
